+++ CIAM/outputHelper.cpp
@@ -47,7 +47,7 @@ void fileoutput3( string var1name,string var2name,string var3name,
 			  string var4name,string var5name,string uname,vector<double> dout) {
 	outfile <<var1name<<","<<var2name<<","<<var3name<<","
 			<<var4name<<","<<var5name<<","<<uname<<",";
-	for (int i=0;i<dout.size();i++) {
+	for (int i=0;i< static_cast<int>( dout.size() );i++) {
 		outfile << dout[i]<<",";
 	}
 	outfile <<ctime(&ltime);
@@ -81,7 +81,7 @@ void dboutput4(string var1name,string var2name,string var3name,string var4name,
 	DBoutrst.SetField(4L, COleVariant(var3name.c_str(), VT_BSTRT));
 	DBoutrst.SetField(5L, COleVariant(var4name.c_str(), VT_BSTRT));
 	DBoutrst.SetField(6L, COleVariant(uname.c_str(), VT_BSTRT));
-	for (i=0;i<dout.size();i++) {
+	for (i=0;i< static_cast<int>( dout.size() );i++) {
 		j = 7+i;
 		DBoutrst.SetField(j, COleVariant(dout[i]));
 	}
